Move copy into kernels::cast. Simplify imports
