+++ candle-metal-kernels/src/lib.rs
@@ -11,8 +11,7 @@ pub use kernels::{
     call_arg_sort, call_binary_contiguous, call_binary_strided, call_cast_contiguous,
     call_cast_strided, call_const_set_contiguous, call_const_set_contiguous_tiled,
     call_const_set_strided, call_mlx_arg_sort, call_mlx_gemm, call_quantized_matmul_mm_t,
-    call_quantized_matmul_mv_t, call_unary_contiguous, call_unary_contiguous_tiled,
-    call_unary_strided, reduce::*, unary, GemmDType, GgmlDType,
+    call_quantized_matmul_mv_t, reduce::*, unary, unary::*, GemmDType, GgmlDType,
 };
 use metal::{
     BlitCommandEncoder, Buffer, CommandQueue, ComputeCommandEncoder, ComputePipeline,
@@ -44,59 +43,6 @@ impl DType {
             Self::F32 => 4,
         }
     }
-}
-
-pub mod copy2d {
-    pub struct Kernel(pub &'static str);
-    pub const FLOAT: Kernel = Kernel("copy2d_f32");
-    pub const HALF: Kernel = Kernel("copy2d_f16");
-    pub const BFLOAT: Kernel = Kernel("copy2d_bf16");
-    pub const I64: Kernel = Kernel("copy2d_i64");
-    pub const U32: Kernel = Kernel("copy2d_u32");
-    pub const U8: Kernel = Kernel("copy2d_u8");
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn call_copy2d(
-    device: &Device,
-    ep: impl EncoderProvider,
-    kernels: &Kernels,
-    name: copy2d::Kernel,
-    input: &Buffer,
-    output: &Buffer,
-    d1: usize,
-    d2: usize,
-    src_s: usize,
-    dst_s: usize,
-    src_o_in_bytes: usize,
-    dst_o_in_bytes: usize,
-) -> Result<(), MetalKernelError> {
-    let pipeline = kernels.load_pipeline(device, Source::Unary, name.0)?;
-    let encoder = ep.encoder();
-    let encoder: &ComputeCommandEncoder = encoder.as_ref();
-    encoder.set_compute_pipeline_state(&pipeline);
-    set_params!(
-        encoder,
-        (
-            d1 as i64,
-            d2 as i64,
-            src_s as i64,
-            dst_s as i64,
-            (input, src_o_in_bytes),
-            (output, dst_o_in_bytes)
-        )
-    );
-
-    let grid_dims = MTLSize {
-        width: d1,
-        height: d2,
-        depth: 1,
-    };
-    let group_dims = get_block_dims(d1, d2, 1);
-    encoder.use_resource(input, MTLResourceUsage::Read);
-    encoder.use_resource(output, MTLResourceUsage::Write);
-    encoder.dispatch_threads(grid_dims, group_dims);
-    Ok(())
 }
 
 #[allow(clippy::too_many_arguments)]
